refactor(web): emergency / persona の hook をドメイン別ファイルに分離する
useDashboard.ts に集約されていた dashboard 用 hook を
hooks/dashboard/ 配下にドメイン別に分割する第一弾。
keys を hooks/dashboard/keys.ts に共通化し、emergency と persona を
useEmergencies / usePersonas に切り出してテストを追加する。
