+++ web/src/hooks/dashboard/keys.ts
@@ -1,0 +1,14 @@
+export const dashboardKeys = {
+  broadcasts: ["dashboard", "broadcasts"] as const,
+  broadcastDetail: (id: string) => ["dashboard", "broadcast", id] as const,
+  personas: ["dashboard", "personas"] as const,
+  emergencies: ["dashboard", "emergencies"] as const,
+  feedbacks: ["dashboard", "feedbacks"] as const,
+  feedbackDetail: (id: string) => ["dashboard", "feedback", id] as const,
+  knowledgeFiles: ["dashboard", "knowledge", "files"] as const,
+  knowledgeUnifiedFiles: ["dashboard", "knowledge", "unified"] as const,
+  knowledgeFile: (key: string) =>
+    ["dashboard", "knowledge", "file", key] as const,
+  polls: ["dashboard", "polls"] as const,
+  pollResults: (id: string) => ["dashboard", "poll", "results", id] as const,
+};
