+++ web/src/hooks/dashboard/usePersonas.ts
@@ -1,0 +1,37 @@
+import {
+  useInfiniteQuery,
+  useMutation,
+  useQueryClient,
+} from "@tanstack/react-query";
+
+import { personaRepository } from "~/lib/api/repository";
+import { dashboardKeys } from "./keys";
+
+export const usePersonas = (limit = 30) =>
+  useInfiniteQuery({
+    queryKey: [...dashboardKeys.personas, limit],
+    queryFn: ({ pageParam }) =>
+      personaRepository.fetchPersonas({ limit, cursor: pageParam }),
+    initialPageParam: undefined as string | undefined,
+    getNextPageParam: (lastPage) => lastPage.nextCursor ?? undefined,
+  });
+
+export const useExtractPersonas = () => {
+  const queryClient = useQueryClient();
+  return useMutation({
+    mutationFn: personaRepository.extractPersonas,
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: dashboardKeys.personas });
+    },
+  });
+};
+
+export const useDeletePersonas = () => {
+  const queryClient = useQueryClient();
+  return useMutation({
+    mutationFn: personaRepository.deleteAllPersonas,
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: dashboardKeys.personas });
+    },
+  });
+};
